backend/requirements.txt
# Core ML Libraries - CPU-friendly for MacBook Air M2 8GB RAM
torch==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2

# Vector Search and Retrieval
faiss-cpu==1.7.4
numpy==1.24.3
scikit-learn==1.3.2

# FastAPI Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Data Processing
pandas==2.1.3
jsonlines==4.0.0

# Evaluation and Metrics
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
tqdm==4.66.1
python-dotenv==1.0.0
